FaceRecogEngine/requirements.txt
appdirs==1.4.3
Django==2.1
djangorestframework==3.8.2
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
pillow==5.1.0
scipy==1.1.0
scikit-learn==0.19.2
request==1.0.2